Split out guid tests to separate test file
